refactor: вынесла тип ILayer в папку interface
