+++ src/entities/LayersItem/index.ts
@@ -1,1 +1,3 @@
 export { LayersItem } from "./ui";
+
+export type { ILayer } from "./interface";
